refactor: single file to multiple modules
