+++ index.js
@@ -3,9 +3,10 @@
 const controlAccess = require('control-access');
 const etag = require('etag');
 const got = require('got');
-const {parseString} = require('xml2js');
 
-const transformBook = require('./transform-book');
+const fetchReviews = require('./lib/fetch-reviews');
+const reviewToIsbn = require('./lib/review-to-isbn');
+const transformBook = require('./lib/transform-book');
 
 const cache = `max-age=${Number(process.env.CACHE_MAX_AGE) || 300}`;
 const MAX_BOOKS = Number(process.env.MAX_BOOKS) || 10;
@@ -16,7 +17,6 @@ const {
   GOOGLE_BOOKS_API_KEY
 } = process.env;
 
-const isString = subject => typeof subject === 'string';
 const ONE_DAY = 1000 * 60 * 60 * 24;
 
 if (!GOODREADS_API_KEY) {
@@ -40,33 +40,10 @@ let responseETag = '';
 
 const GOOGLE_API = `https://www.googleapis.com/books/v1/volumes?q=isbn:{isbn}&key=${GOOGLE_BOOKS_API_KEY}`;
 
-async function fetchLatest() {
+async function getRecentlyReadBooks() {
   try {
-    const {body} = await got(`https://www.goodreads.com/review/list/${GOODREADS_LIST_ID}.xml?key=${GOODREADS_API_KEY}&v=2&shelf=read&sort=date_read`);
-
-    const reviews = await new Promise((resolve, reject) => {
-      parseString(body, (error, result) => {
-        if (error) {
-          reject(error);
-        }
-
-        const reviews = result.GoodreadsResponse.reviews[0].review;
-        resolve(reviews);
-      });
-    });
-
-    const isbns = reviews
-      .filter(({read_at: readAt}) => {
-        const [date] = readAt;
-        return isString(date) && date.length > 3;
-      })
-      .map(({book}) => {
-        const [{isbn: isbnArr = [], isbn13: isbn13Arr = []}] = book;
-        const [isbn] = isbnArr;
-        const [isbn13] = isbn13Arr;
-        return isString(isbn) ? isbn : isString(isbn13) ? isbn13 : false;
-      })
-      .filter(isbn => isString(isbn));
+    const reviews = await fetchReviews(GOODREADS_API_KEY, GOODREADS_LIST_ID);
+    const isbns = reviews.reduce(reviewToIsbn);
 
     const bookPromises = isbns.map(isbn => got(GOOGLE_API.replace('{isbn}', isbn)));
     const bookResults = await Promise.all(bookPromises);
@@ -79,11 +56,12 @@ async function fetchLatest() {
   }
 }
 
-setInterval(fetchLatest, ONE_DAY);
-fetchLatest();
+setInterval(getRecentlyReadBooks, ONE_DAY);
+getRecentlyReadBooks();
 
 module.exports = (request, response) => {
   controlAccess()(request, response);
+
   response.setHeader('cache-control', cache);
   response.setHeader('etag', responseETag);
