+++ lib/transform-book.js
@@ -1,10 +1,12 @@
+'use strict';
+
 const convertToHttps = require('to-https');
 
 module.exports = result => {
   const {body} = result;
-  const {items = []} = JSON.parse(body);
+  const {items} = JSON.parse(body);
 
-  if (!items.length > 0) {
+  if (!items) {
     return;
   }
